server/requirement.txt
bidict==0.22.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
eventlet==0.33.1
Flask==2.2.2
Flask-SocketIO==5.3.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
python-engineio==4.3.4
python-socketio==5.7.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0